Refactoring and began work on target
Hasher has been split up into Hasher and Validator.

Additionally, a target field was added to Block, with relevent getter
and setter added. It must be specified in the constructor as well.
Hasher now takes target into account when generating the block hash. The
only thing left to add with target is a function to determine what the
target should be.